Added: fragments, changed MAIN & LAUNCHER from MainActivity to SelectCityActivity
