+++ app/src/main/java/com/antonageev/weatherapp/Parcel.java
@@ -1,0 +1,16 @@
+package com.antonageev.weatherapp;
+
+import java.io.Serializable;
+import java.util.Map;
+
+class Parcel implements Serializable {
+    private Map<String, String> mapData;
+
+    Parcel(Map<String, String> data){
+        this.mapData = data;
+    }
+
+    Map<String, String> getMapData() {
+        return mapData;
+    }
+}
