+++ app/src/main/res/layout-land/activity_select_city.xml
