+++ app/src/main/res/layout/activity_settings.xml
@@ -1,129 +1,20 @@
 <?xml version="1.0" encoding="utf-8"?>
-
-
-<androidx.constraintlayout.widget.ConstraintLayout
-    xmlns:android="http://schemas.android.com/apk/res/android"
+<LinearLayout xmlns:android="http://schemas.android.com/apk/res/android"
+    xmlns:app="http://schemas.android.com/apk/res-auto"
+    xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    xmlns:app="http://schemas.android.com/apk/res-auto">
+    >
 
-
-    <Button
-        android:id="@+id/backButton"
+    <FrameLayout
+        android:id="@+id/settingsFragmentContainer"
         android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:text="@string/backButton"
-        app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toTopOf="parent" />
-
-    <TextView
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:text="@string/settings"
-        android:textSize="24sp"
-        app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintBaseline_toBaselineOf="@id/backButton" />
-
-    <TextView
-        android:id="@+id/settingsDarkTheme"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:layout_marginStart="16dp"
-        android:layout_marginTop="16dp"
-        android:text="@string/textDarkTheme"
-        android:textSize="24sp"
-        app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toBottomOf="@id/backButton" />
-
-    <Switch
-        android:id="@+id/switchDarkTheme"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:layout_marginEnd="16dp"
-        android:checked="false"
-        android:text="@string/textSwitchDarkThemeOff"
-        app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintTop_toTopOf="@id/settingsDarkTheme"/>
-
-    <View
-        android:id="@+id/settingsDivider"
-        android:layout_width="match_parent"
-        android:layout_height="1dp"
-        android:layout_marginTop="8dp"
-        android:background="?android:attr/listDivider"
-        app:layout_constraintTop_toBottomOf="@id/switchDarkTheme" />
-
-    <TextView
-        android:id="@+id/settingsWindSpeed"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:layout_marginStart="16dp"
-        android:layout_marginTop="8dp"
-        android:text="@string/wind_speed"
-        android:textSize="24sp"
-        app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toBottomOf="@id/settingsDivider" />
-
-    <RadioGroup
-        android:layout_width="150dp"
-        android:layout_height="wrap_content"
-        android:layout_marginEnd="16dp"
-        android:orientation="horizontal"
-        app:layout_constraintTop_toTopOf="@id/settingsWindSpeed"
-        app:layout_constraintEnd_toEndOf="parent">
-
-        <RadioButton
-            android:id="@+id/radioButtonMS"
-            android:layout_width="75dp"
+        android:layout_height="wrap_content">
+        <fragment
+            android:name="com.antonageev.weatherapp.SettingsFragment"            android:id="@+id/selectCityFragment"
+            android:background="@drawable/cloudy_back"
+            android:layout_width="wrap_content"
             android:layout_height="wrap_content"
-            android:layout_weight="1"
-            android:text="@string/metersPerSecond"
-            android:checked="true"/>
-
-        <RadioButton
-            android:id="@+id/radioButtonKmHour"
-            android:layout_width="75dp"
-            android:layout_height="wrap_content"
-            android:layout_weight="1"
-            android:text="@string/kilometers" />
-
-    </RadioGroup>
-
-    <TextView
-        android:id="@+id/settingsTemperature"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:layout_marginStart="16dp"
-        android:layout_marginTop="16dp"
-        android:text="@string/settingsTemperature"
-        android:textSize="24sp"
-        app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toBottomOf="@id/settingsWindSpeed" />
-
-    <RadioGroup
-        android:layout_width="150dp"
-        android:layout_height="wrap_content"
-        android:layout_marginEnd="16dp"
-        android:orientation="horizontal"
-        app:layout_constraintTop_toTopOf="@id/settingsTemperature"
-        app:layout_constraintEnd_toEndOf="parent">
-
-        <RadioButton
-            android:id="@+id/radioButtonCelsius"
-            android:layout_width="75dp"
-            android:layout_height="wrap_content"
-            android:layout_weight="1"
-            android:text="@string/settingsTextCelsius"
-            android:checked="true"/>
-
-        <RadioButton
-            android:id="@+id/radioButtonFahrenheit"
-            android:layout_width="75dp"
-            android:layout_height="wrap_content"
-            android:layout_weight="1"
-            android:text="@string/settingsTextFahrenheit" />
-
-    </RadioGroup>
-
-</androidx.constraintlayout.widget.ConstraintLayout>
+            tools:layout="@layout/fragment_settings" />
+    </FrameLayout>
+</LinearLayout>
