+++ app/src/main/res/values/strings.xml
@@ -38,5 +38,4 @@
     <string name="urlNewYork">https://en.wikipedia.org/wiki/New_York_City</string>
     <string name="urlTokyo">https://en.wikipedia.org/wiki/Tokyo</string>
     <string name="urlMoscow">https://en.wikipedia.org/wiki/Moscow</string>
-
 </resources>
